♻️ refactor : 즐겨찾기 기능 분리
