+++ src/main/java/nextstep/favorite/domain/FavoriteRepository.java
@@ -1,0 +1,9 @@
+package nextstep.favorite.domain;
+
+import java.util.List;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface FavoriteRepository extends JpaRepository<Favorite, Long> {
+
+    List<Favorite> findByMemberId(Long memberId);
+}
