+++ src/main/java/nextstep/member/domain/Member.java
@@ -4,7 +4,6 @@ import java.util.List;
 import javax.persistence.CollectionTable;
 import javax.persistence.Column;
 import javax.persistence.ElementCollection;
-import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
@@ -29,9 +28,6 @@ public class Member {
     )
     @Column(name = "role")
     private List<String> roles;
-
-    @Embedded
-    private Favorites favorites = new Favorites();
 
     public Member() {
     }
@@ -80,17 +76,5 @@ public class Member {
         if (!this.password.equals(password)) {
             throw new PasswordMismatchException();
         }
-    }
-
-    public void addFavorite(final Favorite favorite) {
-        this.favorites.add(favorite);
-    }
-
-    public List<Favorite> getFavorites() {
-        return favorites.getFavorites();
-    }
-
-    public void deleteFavorite(final Long id) {
-        favorites.delete(id);
     }
 }
